+++ Tools/KFModTool/map.h
@@ -4,7 +4,7 @@
 #include "checksum.h"
 #include "kfmterror.h"
 #include "kftypes.h"
-#include "tfile.h"
+#include "../common/tfile.h"
 #include <QMessageBox>
 
 typedef quint8 byte;
