--- Tools/KFModTool/prettynamer.cpp
@@ -1,1 +1,0 @@
-../tfile_tool/prettynamer.cpp
