--- Tools/KFModTool/prettynamer.h
@@ -1,1 +1,0 @@
-../tfile_tool/prettynamer.h
