--- Tools/KFModTool/tfile.cpp
@@ -1,1 +1,0 @@
-../tfile_tool/tfile.cpp
