--- Tools/KFModTool/tfile.h
@@ -1,1 +1,0 @@
-../tfile_tool/tfile.h
